Moved the http api to its own crate
This allows custom mmids applications that don't want the built in HTTP api to not deal with the dependencies for them.
