+++ mmids-http-api/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "mmids-http-api"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+mmids-core = {path = "../mmids-core"}
+
+async-trait = "0.1"
+bytes = "1.1.0"
+hyper = { version = "0.14", features = ["server", "http2"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+thiserror = "1.0"
+tokio = { version = "1.24", features = ["sync"] }
+tracing = { version = "0.1", features = ["log"] }
+uuid = { version = "1.2", features = ["v4"] }
