+++ mmids-http-api/src/handlers/stop_workflow.rs
@@ -1,9 +1,9 @@
 //! Handler that allows a workflow to be stopped
 
-use crate::http_api::routing::RouteHandler;
-use crate::workflows::manager::{WorkflowManagerRequest, WorkflowManagerRequestOperation};
+use crate::routing::RouteHandler;
 use async_trait::async_trait;
 use hyper::{Body, Error, Request, Response, StatusCode};
+use mmids_core::workflows::manager::{WorkflowManagerRequest, WorkflowManagerRequestOperation};
 use std::collections::HashMap;
 use std::sync::Arc;
 use tokio::sync::mpsc::UnboundedSender;
